test: deflake watch mode tests
PR-URL: https://github.com/nodejs/node/pull/44621
Backport-PR-URL: https://github.com/nodejs/node/pull/44976
Fixes: https://github.com/nodejs/node/issues/44655
