--- test/fixtures/watch-mode/graceful-sigterm.js
@@ -1,17 +1,0 @@
-
-setInterval(() => {}, 1000);
-console.log('running');
-
-process.on('SIGTERM', () => {
-  setTimeout(() => {
-    console.log('exiting gracefully');
-    process.exit(0);
-  }, 1000);
-});
-
-process.on('SIGINT', () => {
-  setTimeout(() => {
-    console.log('exiting gracefully');
-    process.exit(0);
-  }, 1000);
-});
