--- test/fixtures/watch-mode/infinite-loop.js
@@ -1,2 +1,0 @@
-console.log('running');
-while(true) {};
